requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai==1.12.0
python-dotenv==1.0.0
Pillow==10.1.0
requests==2.31.0
pydantic==2.5.2
httpx==0.26.0 